return search_deep_keys to data_structures_utils.py
